requirements.txt
alembic==1.18.2
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysqlclient==2.2.7
passlib==1.7.4
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
